Fix display id type in long press affordance handler. Currently, the affordance does not show up because the display id gets mangled due to mismatching type.
BUG=155523


Review URL: https://chromiumcodereview.appspot.com/11189005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@162219 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 344aa38c679221826d5fb0cf1b04072a818e634d

+++ ash/wm/gestures/long_press_affordance_handler.h
@@ -68,7 +68,7 @@ class LongPressAffordanceHandler : public ui::AnimationDelegate,
   gfx::Point tap_down_location_;
   int tap_down_touch_id_;
   base::OneShotTimer<LongPressAffordanceHandler> timer_;
-  int tap_down_display_id_;
+  int64 tap_down_display_id_;
   LongPressAnimationType current_animation_type_;
 
   DISALLOW_COPY_AND_ASSIGN(LongPressAffordanceHandler);
